(svn r24908) -Fix [FS#5419]: Allow GSs to pass negative integer string parameters (frosch123)

+++ src/script/api/script_text.cpp
@@ -183,7 +183,7 @@ char *ScriptText::_GetEncodedText(char *p, char *lastofp)
 			p = this->paramt[i]->_GetEncodedText(p, lastofp);
 			continue;
 		}
-		p += seprintf(p, lastofp,":%X", this->parami[i]);
+		p += seprintf(p, lastofp,":%X", (uint32)this->parami[i]);
 	}
 
 	return p;
